Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfminer.six==20221105
sentence-transformers==2.6.1
qdrant-client==1.9.1
tqdm==4.66.4